fix bug in OptionMenu.setitems, introduced with the fixing of this python bug: http://bugs.python.org/issue1342811
versions of python 2.5.4 and newer now automatically delete commands when items are deleted
however, that causes error if the command is deleted manually beforehand.
old versions, on the contrary, require manual command deletion to free up the objects.

so we put in a python version check to see if we need to manually delete commands.

fixes the Pmw bug referred to here:
http://sourceforge.net/forum/forum.php?thread_id=3283195&forum_id=33675
and here:
http://sourceforge.net/tracker/?func=detail&aid=2795731&group_id=10743&atid=110743